refactored files to be more organized
