--- src/Basic_Question.css
@@ -1,153 +1,0 @@
-
-/* Header */
-.questionHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #cecece;
-}
-
-/* Dropdown Settings Menu */
-.dropdown {
-  position: relative;
-  display: inline-block;
-  margin-left: 5px;
-  margin-right: 5px;
-}
-
-.dropdownButton {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 8px;
-}
-
-.bar {
-  width: 25px;
-  height: 3px;
-  background-color: rgb(0, 0, 0);
-  margin: 5px 0;
-}
-
-.dropdownContent {
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 100px;
-  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-  right: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-}
-
-.dropdownContent button {
-  display: block;
-  padding: 5px;
-  width: 100%;
-  text-align: left;
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-}
-
-.dropdown.open .dropdownContent {
-  display: block;
-}
-
-.dropdownItem:hover{
-  background-color: #888888;
-}
-
-/* Header Question Counter */
-.questionCounter{
-  margin: 10px;
-}
-
-/* Header Buttons */
-.buttonContainer {
-  display: flex;
-  align-items: center;
-}
-
-.homeButton,
-.saveButton{
-  background-color: #000000;
-  color: rgb(255, 255, 255);
-  border: none;
-  padding: 8px 16px;
-  border-radius: 4px;
-  margin-left: 5px;
-  cursor: pointer;
-  display: flex;
-}
-
-/* Question Answer Buttons */
-.buttonGrid {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 10px;
-}
-
-.bigButton {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  padding: 20px;
-  font-size: 18px;
-}
-
-.bigButton img {
-  width: 300px;
-  margin-bottom: 10px;
-}
-
-/* Detailed Answer Textbox */
-.detailedAnswerText{
-  width: 50%;
-  padding: 20px;
-  height: 300%;
-}
-
-/* Dark mode styles */
-
-.dark-mode .detailedAnswerText{
-  background-color: #121212;
-  color:#ffffff
-}
-
-.dark-mode .container {
-  background-color: #121212;
-  color: #ffffff;
-}
-
-.dark-mode .questionHeader {
-  background-color: #121212;
-  color: #ffffff;
-}
-
-.dark-mode .bar {
-  background-color: white;
-}
-
-.dark-mode .dropdownContent {
-  background-color: #252525;
-  box-shadow: 0px 8px 16px 0px rgba(255,255,255,0.2);
-}
-
-.dark-mode .dropdownContent button {
-  color: #ffffff;
-}
-
-.dark-mode .homeButton,
-.dark-mode .saveButton {
-  background-color: #ffffff;
-  color: #000000;
-}
-
-.dark-mode .bigButton {
-  background-color: #252525;
-  color: #ffffff;
-}
